requirements.txt
matplotlib==1.5.3
numpy==1.12.1
Pillow==4.0.0
protobuf==3.2.0
PyQt5==5.7.1
sip==4.19
six==1.10.0
tensorflow==1.1.0